requirements.txt
Flask-WTF==0.14.2
PyMySQL==0.7.11
WTForms-Components==0.10.3